LayoutTests/http/tests/appcache/wrong-signature.html

<html manifest="resources/wrong-content-type.manifest">
<body>
<p>Test that a manifest served with a wrong signature isn't treated as such.</p>
<p>Should say SUCCESS:</p>
<div id=result></div>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText()
    layoutTestController.waitUntilDone();
}

function log(message)
{
    document.getElementById("result").innerHTML += message + "<br>";
}

function cacheCallback()
{
    log("FAIL: An event was dispatched on window.applicationCache, even though it was not associated wuth any cache yet.");
}

function timeout()
{
    // Wait for cache update to finish.
    if (window.applicationCache.status != 0) {
        setTimeout(timeout, 0);
        return;
    }
    
    log("SUCCESS");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

applicationCache.addEventListener('cached', function() { log("cached"); cacheCallback() }, false);
applicationCache.addEventListener('noupdate', function() { log("noupdate"); cacheCallback() }, false);
applicationCache.addEventListener('error', function() { log("error"); cacheCallback() }, false);

setTimeout(timeout, 100);

</script>
</body>
</html>
